requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
librosa
soundfile
numpy
scipy
joblib
xgboost
torch
torchaudio
speechbrain
resemblyzer
python-telegram-bot
